// src/Page/PokeInfo/types.d.ts
  export default interface Type {
    slot: number
    type: Type2
  }
  
  export interface Type2 {
    name: string
    url: string
  }